Log out moved to Common class
