+++ app/src/main/java/com/example/salonappnew/about.java
@@ -14,6 +14,8 @@ import androidx.drawerlayout.widget.DrawerLayout;
 import android.os.Bundle;
 import android.view.View;
 
+import com.example.salonappnew.ui.Dashboard;
+
 public class about extends AppCompatActivity {
     DrawerLayout drawerLayout;
 
@@ -31,13 +33,13 @@ public class about extends AppCompatActivity {
         drawer.closeDrawer(drawerLayout);
     }
     public void ClickHome(View view){
-        drawer.redirectActivity(this,drawer.class);
+        drawer.redirectActivity(this, Dashboard.class);
     }
     public void ClickDashboard(View view){
-        drawer.redirectActivity(this,home.class);
+        drawer.redirectActivity(this,Dashboard.class);
     }
     public void ClickAboutUs(View view){
-        recreate();
+
     }
     public void ClickLogout(View view){
         drawer.logout(this);
